+++ example/sovetromantica.dart
@@ -1,0 +1,10 @@
+import 'package:libanime/libanime.dart';
+
+/// SovetRomantica Example
+void main(List<String> args) async {
+  final sr = SovetRomantica();
+
+  final video = await sr
+      .parse("https://sovetromantica.com/embed/episode_1459_1-subtitles");
+  print(video.url);
+}