+++ lib/libanime.dart
@@ -4,4 +4,6 @@ export './parsers/anime/csst/csst.dart' show CSST;
 export './parsers/anime/kodik/kodik.dart' show Kodik;
 export './parsers/anime/sibnet/sibnet.dart' show Sibnet;
 export './utils/aniskip.dart' show AniSkip;
-export './parsers/manga/mangalib/mangalib.dart' show Mangalib;
+export './parsers/manga/mangalib/mangalib.dart' hide Mangalib;
+export './parsers/anime/sovetromantica/sovetromantica.dart' show SovetRomantica;
+export './parsers/anime/aniboom/aniboom.dart' show Aniboom;