+++ lib/parsers/detect_classes.dart
@@ -1,0 +1,10 @@
+import 'package:libanime/libanime.dart';
+import 'package:libanime/parsers/anime/aniboom/aniboom.dart';
+
+Map<String, dynamic> classes = {
+  "kodik": Kodik(),
+  "csst": CSST(),
+  "sibnet": Sibnet(),
+  "sovetromantica": SovetRomantica(),
+  "aniboom": Aniboom()
+};