+++ lib/wrappers/anime/anime365/endpoints.dart
@@ -1,0 +1,7 @@
+// Endpoints list
+// Main URL
+const mainurl = 'https://smotret-anime.online/api/';
+
+// Series and filters
+const series = 'series';
+const malIdFilter = 'myAnimeListId=';